requirements.txt
flask==3.0.2
pymongo==4.6.1
python-dotenv==1.0.1
openai==1.12.0
bcrypt==4.1.2
pytz==2024.1
flask-session==0.8.0 